v0.0.2
A profound update of algorithms, which corrects allele names when the
same consensus sequence is aligned to multiple reference alleles in
different strains. This problem caused split of allele calls when they
actually reveal the same allele. This update also enables PAMmaker to
take as input the clustering outcome when consensus sequences are
clustered at a nucleotide identity of <100%.